evo2-sae: move steering-analysis harness into evo2_sae.eval.steering
Relocate the steering dose-response / selectivity metrics from evo2_sae.steer_analysis into the
evo2_sae.eval package (src/evo2_sae/steer_analysis.py -> src/evo2_sae/eval/steering.py), alongside
the eval/probing harness. Update the importers (scripts/steer.py CLI + tests/test_steer_analysis.py)
to evo2_sae.eval.steering. The CI lane is dropped via the rebase onto the updated #1622. Pure-CPU
tests, no GPU/model.
